% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyexp-experiment.R
\name{autoplot.glyexp_experiment}
\alias{autoplot.glyexp_experiment}
\alias{autoplot.SummarizedExperiment}
\title{Plots for Glycoverse Data Containers}
\usage{
\method{autoplot}{glyexp_experiment}(object, type = "heatmap", group_col = "group", ...)

\method{autoplot}{SummarizedExperiment}(object, type = "heatmap", group_col = "group", ...)
}
\arguments{
\item{object}{A supported Glycoverse data container.}

\item{type}{The type of plot, one of "heatmap" (default) or "boxplot".}

\item{group_col}{A character string specifying where to find the group information.
It should be a column in the sample information tibble. Defaults to "group".
Only applicable to "boxplot".}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}}, \code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}},
and other compatible \code{SummarizedExperiment} objects.
Possible \code{type}s of plots:
\itemize{
\item "heatmap": (Default) Expression heatmap with columns as samples and rows as variables.
\item "boxplot": Boxplot of expression values for each sample.
}
}
