% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-dea.R
\name{autoplot.glystats_anova_res}
\alias{autoplot.glystats_anova_res}
\alias{autoplot.glystats_kruskal_res}
\title{Plots for Multi-Group Differential Expression Analysis (DEA)}
\usage{
\method{autoplot}{glystats_anova_res}(object, p_cutoff = 0.05, p_col = "p_adj", ...)

\method{autoplot}{glystats_kruskal_res}(object, p_cutoff = 0.05, p_col = "p_adj", ...)
}
\arguments{
\item{object}{A \code{glystats_anova_res} or \code{glystats_kruskal_res} object.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p" (raw p-values without multiple testing correction).}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{glystats_anova_res} and \code{glystats_kruskal_res}.
Draw a dotchart for p-values.
}
