% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-cc.R
\name{autoplot.glystats_cc_res}
\alias{autoplot.glystats_cc_res}
\title{Plots for Consensus Clustering}
\usage{
\method{autoplot}{glystats_cc_res}(object, ...)
}
\arguments{
\item{object}{A \code{glystats_cc_res} object.}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Raise an error for now.
Suggest users to use \code{output_file} parameter in \code{gly_consensus_clustering()} instead.
}
