% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-cor.R
\name{autoplot.glystats_cor_res}
\alias{autoplot.glystats_cor_res}
\title{Plots for Correlation Analysis}
\usage{
\method{autoplot}{glystats_cor_res}(object, ...)
}
\arguments{
\item{object}{A \code{glystats_cor_res} object.}

\item{...}{Other arguments passed to \code{\link[GGally:ggcorr]{GGally::ggcorr()}}.}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{\link[glystats:gly_cor]{glystats::gly_cor()}} results (\code{glystats_cor_res} objects).
}
