% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-cox.R
\name{autoplot.glystats_cox_res}
\alias{autoplot.glystats_cox_res}
\title{Plots for Cox Proportional Hazards Model}
\usage{
\method{autoplot}{glystats_cox_res}(object, p_cutoff = 0.05, p_col = "p_adj", ...)
}
\arguments{
\item{object}{A \code{glystats_cox_res} object.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p_val" (raw p-values without multiple testing correction).}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_cox]{glystats::gly_cox()}} (\code{glystats_cox_res} objects).
Draw a dotchart for p-values.
}
