% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-dea.R
\name{autoplot.glystats_dea_res_anova}
\alias{autoplot.glystats_dea_res_anova}
\alias{autoplot.glystats_dea_res_kruskal}
\title{Plots for Multi-Group Differential Expression Analysis (DEA)}
\usage{
\method{autoplot}{glystats_dea_res_anova}(object, p_cutoff = 0.05, p_col = "p_adj", sig_color = "#FF7777", ...)

\method{autoplot}{glystats_dea_res_kruskal}(object, p_cutoff = 0.05, p_col = "p_adj", sig_color = "#FF7777", ...)
}
\arguments{
\item{object}{A \code{glystats_dea_res_anova} or \code{glystats_dea_res_kruskal} object.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p" (raw p-values without multiple testing correction).}

\item{sig_color}{The color for significant candidates. Defaults to "#FF7777".}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{glystats_dea_res_anova} and \code{glystats_dea_res_kruskal}.
Draw a lolipop plot for p-values.
}
