% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-dea.R
\name{autoplot.glystats_dea_res_ttest}
\alias{autoplot.glystats_dea_res_ttest}
\alias{autoplot.glystats_dea_res_wilcoxon}
\title{Plots for 2-Group Differential Expression Analysis (DEA)}
\usage{
\method{autoplot}{glystats_dea_res_ttest}(
  object,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  up_color = "#FF7777",
  down_color = "#7DA8E6",
  ...
)

\method{autoplot}{glystats_dea_res_wilcoxon}(
  object,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  up_color = "#FF7777",
  down_color = "#7DA8E6",
  ...
)
}
\arguments{
\item{object}{A \code{glystats_dea_res_ttest} or \code{glystats_dea_res_wilcoxon} object.}

\item{log2fc_cutoff}{The log2 fold change cutoff. Defaults to 1.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p" (raw p-values without multiple testing correction).}

\item{up_color}{The color for up-regulated candidates. Defaults to "#FF7777".}

\item{down_color}{The color for down-regulated candidates. Defaults to "#7DA8E6".}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{glystats_dea_res_ttest} and \code{glystats_dea_res_wilcoxon}.
Draw a volcano plot.
}
