% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-enrich.R
\name{autoplot.glystats_go_ora_res}
\alias{autoplot.glystats_go_ora_res}
\alias{autoplot.glystats_kegg_ora_res}
\alias{autoplot.glystats_reactome_ora_res}
\title{Plots for Enrichment Analysis}
\usage{
\method{autoplot}{glystats_go_ora_res}(object, type = "dotplot", ...)

\method{autoplot}{glystats_kegg_ora_res}(object, type = "dotplot", ...)

\method{autoplot}{glystats_reactome_ora_res}(object, type = "dotplot", ...)
}
\arguments{
\item{object}{A \code{glystats_go_ora_res}, \code{glystats_kegg_ora_res}, or \code{glystats_reactome_ora_res} object.}

\item{type}{The type of plot, one of "dotplot" (default), "barplot", or "network".}

\item{...}{Other arguments passed to underlying functions
(\code{\link[enrichplot:dotplot]{enrichplot::dotplot()}} for "dotplot", \code{\link[enrichplot:barplot.enrichResult]{enrichplot::barplot.enrichResult()}} for "barplot", and \code{\link[enrichplot:emapplot]{enrichplot::emapplot()}} for "network")}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_go()}}, \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}, and \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_reactome()}}
(\code{glystats_go_ora_res}, \code{glystats_kegg_ora_res}, and \code{glystats_reactome_ora_res} objects).
Possible \code{type}s of plots:
\itemize{
\item "dotplot" (default): Dotplot showing p-value, gene counts, and gene ratios.
\item "barplot": Barplot showing p-values and gene counts.
\item "network": Network plot showing terms and their relationships.
}
}
\seealso{
\code{\link[enrichplot:dotplot]{enrichplot::dotplot()}}, \code{\link[enrichplot:barplot.enrichResult]{enrichplot::barplot.enrichResult()}}, \code{\link[enrichplot:emapplot]{enrichplot::emapplot()}}
}
