% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-hclust.R
\name{autoplot.glystats_hclust_res}
\alias{autoplot.glystats_hclust_res}
\title{Plots for Hierarchical Clustering}
\usage{
\method{autoplot}{glystats_hclust_res}(object, ...)
}
\arguments{
\item{object}{A \code{glystats_hclust_res} object.}

\item{...}{Additional arguments passed to \code{\link[ggdendro:ggdendrogram]{ggdendro::ggdendrogram()}}.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_hclust]{glystats::gly_hclust()}} (\code{glystats_hclust_res} object).
Draw a dendrogram.
}
\seealso{
\code{\link[ggdendro:ggdendrogram]{ggdendro::ggdendrogram()}}
}
