% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-plsda.R
\name{autoplot.glystats_plsda_res}
\alias{autoplot.glystats_plsda_res}
\title{Plots for PLS-DA}
\usage{
\method{autoplot}{glystats_plsda_res}(object, type = "scores", groups = NULL, group_col = NULL, ...)
}
\arguments{
\item{object}{A \code{glystats_plsda_res} object.}

\item{type}{The type of plot, one of "loadings", "scores", "vip", or "variance". Defaults to "scores".}

\item{groups}{A factor or character vector specifying group membership for each sample.
If provided, the plot will be colored by group.
Only applicable to "scores" and "loadings" types.}

\item{group_col}{A character string specifying where to find the group information.
If you uses \code{\link[glystats:gly_plsda]{glystats::gly_plsda()}} on a \code{\link[glyexp:experiment]{glyexp::experiment()}} to get the result,
sample information has already been added to the result.
In this case, you can specify the column name in the sample information tibble
to be used for coloring.
If not provided, this function will try "group".}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_plsda]{glystats::gly_plsda()}} (\code{glystats_plsda_res} object).
Possible \code{type}s of plots:
\itemize{
\item "scores" (default): Plot scores for samples.
\item "loadings": Plot loadings for variables.
\item "vip": Plot VIP scores for variables.
\item "variance": Plot explained variance for each component.
}
}
