% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-roc.R
\name{autoplot.glystats_roc_res}
\alias{autoplot.glystats_roc_res}
\title{Plots for ROC Analysis}
\usage{
\method{autoplot}{glystats_roc_res}(object, auc_cutoff = 0.5, ...)
}
\arguments{
\item{object}{A \code{glystats_roc_res} object.}

\item{auc_cutoff}{The AUC cutoff. Defaults to 0.5.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_roc]{glystats::gly_roc()}} (\code{glystats_roc_res} objects).
Draw a dotchart for AUC values.
}
