% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-dea.R
\name{autoplot.glystats_ttest_res}
\alias{autoplot.glystats_ttest_res}
\alias{autoplot.glystats_wilcox_res}
\title{Plots for 2-Group Differential Expression Analysis (DEA)}
\usage{
\method{autoplot}{glystats_ttest_res}(object, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

\method{autoplot}{glystats_wilcox_res}(object, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)
}
\arguments{
\item{object}{A \code{glystats_ttest_res} or \code{glystats_wilcox_res} object.}

\item{log2fc_cutoff}{The log2 fold change cutoff. Defaults to 1.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p_val" (raw p-values without multiple testing correction).}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}} and \code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}}
(\code{glystats_ttest_res} and \code{glystats_wilcox_res} objects).
Draw a volcano plot.
}
