% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-umap.R
\name{autoplot.glystats_umap_res}
\alias{autoplot.glystats_umap_res}
\title{Plots for UMAP}
\usage{
\method{autoplot}{glystats_umap_res}(object, groups = NULL, group_col = NULL, ...)
}
\arguments{
\item{object}{A \code{glystats_umap_res} object.}

\item{groups}{A factor or character vector specifying group membership for each sample.
If provided, the plot will be colored by group.}

\item{group_col}{A character string specifying where to find the group information.
If you uses \code{\link[glystats:gly_umap]{glystats::gly_umap()}} on a \code{\link[glyexp:experiment]{glyexp::experiment()}} to get the result,
sample information has already been added to the result.
In this case, you can specify the column name in the sample information tibble
to be used for coloring.
If not provided, this function will try "group".}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_umap]{glystats::gly_umap()}} (\code{glystats_umap_res} objects).
Draw a scatter plot of UMAP coordinates.
}
