% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-wgcna.R
\name{autoplot.glystats_wgcna_res}
\alias{autoplot.glystats_wgcna_res}
\title{Plots for WGCNA}
\usage{
\method{autoplot}{glystats_wgcna_res}(object, ...)
}
\arguments{
\item{object}{A \code{glystats_wgcna_res} object.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_wgcna]{glystats::gly_wgcna()}} (\code{glystats_wgcna_res} objects).
Draw a heatmap of eigenvalues.
}
