% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyvis-package.R
\docType{package}
\name{glyvis-package}
\alias{glyvis}
\alias{glyvis-package}
\title{glyvis: Fast Visualization For Glycoverse}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a unified interface for visualizing glycomics data within the 'glycoverse' ecosystem. Implements 'autoplot()' methods for various data structures including statistical results from 'glystats', experimental data from 'glyexp', and glycan structures from 'glyrepr'. Designed for rapid data exploration and quality assessment rather than publication-quality figure generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyvis/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
