% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\alias{plot_boxplot.glyexp_experiment}
\alias{plot_boxplot.SummarizedExperiment}
\title{Boxplot}
\usage{
plot_boxplot(x, ...)

\method{plot_boxplot}{glyexp_experiment}(x, group_col = "group", ...)

\method{plot_boxplot}{SummarizedExperiment}(x, group_col = "group", ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Ignored.}

\item{group_col}{A character string specifying the column name in sample information
that contains group labels. Default is "group".}
}
\value{
A ggplot object.
}
\description{
Draw a boxplot.
Currently supported data types:
\itemize{
\item \code{SummarizedExperiment}: A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}},
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or other compatible container.
Boxplots of log2-transformed expression values are plotted, grouped by the group column.
}
}
