% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-enrich.R
\name{plot_enrich}
\alias{plot_enrich}
\alias{plot_enrich.glystats_go_ora_res}
\alias{plot_enrich.glystats_kegg_ora_res}
\alias{plot_enrich.glystats_reactome_ora_res}
\alias{plot_enrich.glyexp_experiment}
\title{Plot Enrichment Analysis Results}
\usage{
plot_enrich(x, type = "dotplot", ...)

\method{plot_enrich}{glystats_go_ora_res}(x, type = "dotplot", ...)

\method{plot_enrich}{glystats_kegg_ora_res}(x, type = "dotplot", ...)

\method{plot_enrich}{glystats_reactome_ora_res}(x, type = "dotplot", ...)

\method{plot_enrich}{glyexp_experiment}(x, type = "dotplot", enrich_type = "go", stats_args = list(), ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{type}{The type of plot, one of "dotplot" (default), "barplot", or "network".}

\item{...}{Additional arguments passed to underlying functions:
\itemize{
\item "dotplot": \code{\link[enrichplot:dotplot]{enrichplot::dotplot()}}
\item "barplot": \code{\link[enrichplot:barplot.enrichResult]{enrichplot::barplot.enrichResult()}}
\item "network": \code{\link[enrichplot:emapplot]{enrichplot::emapplot()}}
}}

\item{enrich_type}{The type of enrichment analysis, one of "go" (default), "kegg", or "reactome".}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_go()}},
\code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}, or \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_reactome()}}.}
}
\value{
A ggplot object.
}
\description{
Draw a enrichment analysis result plot.
Currently supported data types:
\itemize{
\item \code{glystats_go_ora_res}: Result from \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_go()}}.
\item \code{glystats_kegg_ora_res}: Result from \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}}.
\item \code{glystats_reactome_ora_res}: Result from \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_reactome()}}.
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
Enrichment analysis is first performed using \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_go()}}, \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_kegg()}},
or \code{\link[glystats:gly_enrich_go]{glystats::gly_enrich_reactome()}}, then the result is plotted.
}
}
