% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.glyexp_experiment}
\alias{plot_heatmap.SummarizedExperiment}
\title{Heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{glyexp_experiment}(x, ...)

\method{plot_heatmap}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A supported Glycoverse data container.}

\item{...}{Other arguments passed to \code{pheatmap::pheatmap()}.}
}
\value{
A ggplot object.
}
\description{
Draw a heatmap from a \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}},
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or other compatible \code{SummarizedExperiment}.
Heatmap of log2-transformed expression values is plotted.
Before plotting, zero-variance rows and columns are filtered out
to ensure robust clustering.
}
