% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-oplsda.R
\name{plot_oplsda}
\alias{plot_oplsda}
\alias{plot_oplsda.glystats_oplsda_res}
\alias{plot_oplsda.glyexp_experiment}
\title{OPLS-DA Plot}
\usage{
plot_oplsda(
  x,
  type = "scores",
  y_type = "o1",
  groups = NULL,
  group_col = NULL,
  ...
)

\method{plot_oplsda}{glystats_oplsda_res}(
  x,
  type = "scores",
  y_type = "o1",
  groups = NULL,
  group_col = NULL,
  ...
)

\method{plot_oplsda}{glyexp_experiment}(
  x,
  type = "scores",
  y_type = "o1",
  groups = NULL,
  group_col = NULL,
  stats_args = list(),
  ...
)
}
\arguments{
\item{x}{An object to be plotted.}

\item{type}{The type of plot, one of "loadings", "scores", "vip", "variance", or "s-plot". Defaults to "scores".}

\item{y_type}{What to plot on the y-axis when \code{type} is "scores" or "loadings". Either "p2" or "o1". Defaults to "o1".}

\item{groups}{A factor or character vector specifying group membership for each sample.
If provided, the plot will be colored by group.
Only applicable to "scores" and "loadings" types.}

\item{group_col}{A character string specifying where to find the group information.
If you uses \code{\link[glystats:gly_oplsda]{glystats::gly_oplsda()}} on a \code{\link[glyexp:experiment]{glyexp::experiment()}} to get the result,
sample information has already been added to the result.
In this case, you can specify the column name in the sample information tibble
to be used for coloring.
If not provided, this function will try "group".}

\item{...}{Ignored.}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_oplsda]{glystats::gly_oplsda()}}.}
}
\value{
A ggplot object.
}
\description{
Draw a OPLS-DA scores plot.
Currently supported data types:
\itemize{
\item \code{glystats_oplsda_res}: Result from \code{\link[glystats:gly_oplsda]{glystats::gly_oplsda()}}.
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
OPLS-DA analysis is first performed using \code{\link[glystats:gly_oplsda]{glystats::gly_oplsda()}},
then the result is plotted.
}
}
