% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pca.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA Plot}
\usage{
plot_pca(exp)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}
}
\value{
A ggplot object.
}
\description{
This function accepts a \code{\link[glyexp:experiment]{glyexp::experiment()}},
performs PCA using \code{\link[glystats:gly_pca]{glystats::gly_pca()}},
and plots the PCA scores plot.
}
