% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plsda.R
\name{plot_plsda}
\alias{plot_plsda}
\alias{plot_plsda.glystats_plsda_res}
\alias{plot_plsda.glyexp_experiment}
\alias{plot_plsda.SummarizedExperiment}
\title{PLS-DA Plot}
\usage{
plot_plsda(x, type = "scores", groups = NULL, group_col = NULL, ...)

\method{plot_plsda}{glystats_plsda_res}(x, type = "scores", groups = NULL, group_col = NULL, ...)

\method{plot_plsda}{glyexp_experiment}(
  x,
  type = "scores",
  groups = NULL,
  group_col = NULL,
  stats_args = list(),
  ...
)

\method{plot_plsda}{SummarizedExperiment}(
  x,
  type = "scores",
  groups = NULL,
  group_col = NULL,
  stats_args = list(),
  ...
)
}
\arguments{
\item{x}{An object to be plotted.}

\item{type}{The type of plot, one of "loadings", "scores", "vip", or "variance". Defaults to "scores".}

\item{groups}{A factor or character vector specifying group membership for each sample.
If provided, the plot will be colored by group.
Only applicable to "scores" and "loadings" types.}

\item{group_col}{A character string specifying where to find the group information.
If the result was produced from a supported Glycoverse data container,
sample information has already been added to the result.
In this case, you can specify the column name in the sample information tibble
to be used for coloring.
If not provided, this function will try "group".}

\item{...}{Ignored.}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_plsda]{glystats::gly_plsda()}}.}
}
\value{
A ggplot object.
}
\description{
Draw a PLS-DA scores plot.
Currently supported data types:
\itemize{
\item \code{glystats_plsda_res}: Result from \code{\link[glystats:gly_plsda]{glystats::gly_plsda()}}.
\item \code{SummarizedExperiment}: A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}},
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or other compatible container.
PLS-DA analysis is first performed using \code{\link[glystats:gly_plsda]{glystats::gly_plsda()}},
then the result is plotted.
}
}
