% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roc.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC plot}
\usage{
plot_roc(exp, group_col = "group")
}
\arguments{
\item{exp}{A \code{glyexp::experiment()} object with no more than 10 variables.}

\item{group_col}{A character string specifying the column name of the grouping variable.
Defaults to "group".}
}
\value{
A ggplot object.
}
\description{
Plot ROC curves for all variables in an experiment.
Only two groups are allowed.
At most 10 variables can be plotted.
}
