% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{t-SNE Plot}
\usage{
plot_tsne(exp, ...)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{...}{Additional arguments passed to \code{\link[glystats:gly_tsne]{glystats::gly_tsne()}}.}
}
\value{
A ggplot object.
}
\description{
This function accepts a \code{\link[glyexp:experiment]{glyexp::experiment()}},
performs t-SNE using \code{\link[glystats:gly_tsne]{glystats::gly_tsne()}},
and plots the t-SNE scores plot.
}
