% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-umap.R
\name{plot_umap}
\alias{plot_umap}
\alias{plot_umap.glystats_umap_res}
\alias{plot_umap.glyexp_experiment}
\alias{plot_umap.SummarizedExperiment}
\title{UMAP Plot}
\usage{
plot_umap(x, ...)

\method{plot_umap}{glystats_umap_res}(x, groups = NULL, group_col = NULL, ...)

\method{plot_umap}{glyexp_experiment}(x, groups = NULL, group_col = NULL, stats_args = list(), ...)

\method{plot_umap}{SummarizedExperiment}(x, groups = NULL, group_col = NULL, stats_args = list(), ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Ignored.}

\item{groups}{A factor or character vector specifying group membership for each sample.
If provided, the plot will be colored by group.}

\item{group_col}{A character string specifying where to find the group information.
If the result was produced from a supported Glycoverse data container,
sample information has already been added to the result.
In this case, you can specify the column name in the sample information tibble
to be used for coloring.
If not provided, this function will try "group".}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_umap]{glystats::gly_umap()}}.}
}
\value{
A ggplot object.
}
\description{
Draw a UMAP scores plot.
Currently supported data types:
\itemize{
\item \code{glystats_umap_res}: Result from \code{\link[glystats:gly_umap]{glystats::gly_umap()}}.
\item \code{SummarizedExperiment}: A \code{\link[glyexp:GlycomicSE]{glyexp::GlycomicSE()}},
\code{\link[glyexp:GlycoproteomicSE]{glyexp::GlycoproteomicSE()}}, or other compatible container.
UMAP analysis is first performed using \code{\link[glystats:gly_umap]{glystats::gly_umap()}},
then the result is plotted.
}
}
