% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{autoplot.glyexp_experiment}
\alias{autoplot.glyexp_experiment}
\title{Plots for Experiments}
\usage{
\method{autoplot}{glyexp_experiment}(object, type = "barplot", ...)
}
\arguments{
\item{object}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{type}{The type of plot, one of "heatmap" (default) or "barplot".}

\item{...}{Other arguments passed to underlying functions
(\code{\link[tidyplots:add_boxplot]{tidyplots::add_boxplot()}} for barplot, \code{\link[tidyplots:add_heatmap]{tidyplots::add_heatmap()}} for heatmap).}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{\link[glyexp:experiment]{glyexp::experiment()}}.
Possible \code{type}s of plots:
\itemize{
\item "barplot" (default): Barplot of expression values for each sample.
\item "heatmap": Expression heatmap with columns as samples and rows as variables.
}
}
