% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-dea.R
\name{autoplot.glystats_limma_res}
\alias{autoplot.glystats_limma_res}
\title{Plots for Limma Result}
\usage{
\method{autoplot}{glystats_limma_res}(
  object,
  contrast = NULL,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  ...
)
}
\arguments{
\item{object}{A \code{glystats_limma_res} object.}

\item{contrast}{A character string specifying the contrast to plot, in the format of "group1_vs_group2".
Must be one of the contrasts in the result.
When there is only one contrast (two-group comparison), it can be NULL (default).}

\item{log2fc_cutoff}{The log2 fold change cutoff. Defaults to 1.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p_val" (raw p-values without multiple testing correction).}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_limma]{glystats::gly_limma()}} (\code{glystats_limma_res} objects).
Only 2-group comparison is supported, and a volcano plot is drawn.
}
