% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-oplsda.R
\name{autoplot.glystats_oplsda_res}
\alias{autoplot.glystats_oplsda_res}
\title{Plots for OPLS-DA}
\usage{
\method{autoplot}{glystats_oplsda_res}(object, type = "loadings", y_type = "o1", ...)
}
\arguments{
\item{object}{A \code{glystats_oplsda_res} object.}

\item{type}{The type of plot, one of "loadings", "scores", "vip", or "screeplot".}

\item{y_type}{What to plot on the y-axis when \code{type} is "scores" or "loadings". Either "p2" or "o1".}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_oplsda]{glystats::gly_oplsda()}} (\code{glystats_oplsda_res} object).
Possible \code{type}s of plots:
\itemize{
\item "loadings": Plot loadings for variables.
\item "scores": Plot scores for samples.
\item "vip": Plot VIP scores for variables.
\item "screeplot": Plot explained variance for each component.
}
}
