% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-pca.R
\name{autoplot.glystats_pca_res}
\alias{autoplot.glystats_pca_res}
\title{Plots for Principal Component Analysis (PCA)}
\usage{
\method{autoplot}{glystats_pca_res}(object, type = "individual", group_col = "group", ...)
}
\arguments{
\item{object}{A \code{glystats_pca_res} object.}

\item{type}{The type of plot, one of "individual" (default) or "variables".}

\item{group_col}{For "individual" plot, the column name of the group information in the sample information.}

\item{...}{Other arguments passed to underlying functions.}
}
\value{
A ggplot object.
}
\description{
Visualization for \code{glystats_pca_res} results.
Possible \code{type}s of plots:
\itemize{
\item "individual" (default): Plot samples as individuals.
\item "variables": Plot loadings for variables.
}
}
