% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glystats-roc.R
\name{autoplot.glystats_roc_res}
\alias{autoplot.glystats_roc_res}
\title{Plots for ROC Analysis}
\usage{
\method{autoplot}{glystats_roc_res}(object, type = "dotplot", auc_cutoff = 0.5, ...)
}
\arguments{
\item{object}{A \code{glystats_roc_res} object.}

\item{type}{"dotplot" or "roc". Defaults to "dotplot".
If "roc", at most 10 variables can be plotted.}

\item{auc_cutoff}{The AUC cutoff. Defaults to 0.5. Only used if type is "dotplot".}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Visualization for results from \code{\link[glystats:gly_roc]{glystats::gly_roc()}} (\code{glystats_roc_res} objects).
At most 10 variables can be plotted.
}
