% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyvis-package.R
\docType{package}
\name{glyvis-package}
\alias{glyvis}
\alias{glyvis-package}
\title{glyvis: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyvis/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
