% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(exp, group_col = "group")
}
\arguments{
\item{exp}{A \code{glyexp_experiment} object.}

\item{group_col}{A character string specifying the column name in sample information
that contains group labels. Default is "group".}
}
\value{
A ggplot object.
}
\description{
Draw a boxplot from a \code{\link[glyexp:experiment]{glyexp::experiment()}}.
There should be a group column in the sample information tibble.
}
