% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\alias{plot_boxplot.glyexp_experiment}
\title{Boxplot}
\usage{
plot_boxplot(x, ...)

\method{plot_boxplot}{glyexp_experiment}(x, group_col = "group", ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Ignored.}

\item{group_col}{A character string specifying the column name in sample information
that contains group labels. Default is "group".}
}
\value{
A ggplot object.
}
\description{
Draw a boxplot.
}
