% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-corrplot.R
\name{plot_corrplot}
\alias{plot_corrplot}
\alias{plot_corrplot.glystats_cor_res}
\alias{plot_corrplot.glyexp_experiment}
\title{Correlation Plot}
\usage{
plot_corrplot(x, ...)

\method{plot_corrplot}{glystats_cor_res}(x, ...)

\method{plot_corrplot}{glyexp_experiment}(x, stats_args = list(), ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Additional arguments passed to \code{\link[GGally:ggcorr]{GGally::ggcorr()}}.}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_cor]{glystats::gly_cor()}}.}
}
\value{
A ggplot object.
}
\description{
Draw a correlation matrix heatmap.
Currently supported data types:
\itemize{
\item \code{glystats_cor_res}: Result from \code{\link[glystats:gly_cor]{glystats::gly_cor()}}.
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
Correlation analysis is first performed using \code{\link[glystats:gly_cor]{glystats::gly_cor()}},
then the result is plotted.
}
}
