% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-corrplot.R
\name{plot_corrplot}
\alias{plot_corrplot}
\alias{plot_corrplot.glystats_cor_res}
\alias{plot_corrplot.glyexp_experiment}
\title{Correlation Plot}
\usage{
plot_corrplot(x, ...)

\method{plot_corrplot}{glystats_cor_res}(x, ...)

\method{plot_corrplot}{glyexp_experiment}(x, on = "variable", ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Additional arguments passed to \code{\link[GGally:ggcorr]{GGally::ggcorr()}}.}

\item{on}{A character string specifying what to correlate. Either "variable" (default) to correlate
variables/features, or "sample" to correlate samples/observations.}
}
\value{
A ggplot object.
}
\description{
Draw a correlation matrix heatmap.
}
