% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap}
\usage{
plot_heatmap(exp)
}
\arguments{
\item{exp}{A \code{glyexp_experiment} object.}
}
\value{
A ggplot object.
}
\description{
Draw a heatmap from a \code{\link[glyexp:experiment]{glyexp::experiment()}}.
}
