% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.glyexp_experiment}
\title{Heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{glyexp_experiment}(x, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Draw a heatmap.
}
