% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.glyexp_experiment}
\title{Heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{glyexp_experiment}(x, ...)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{...}{Other arguments passed to \code{pheatmap::pheatmap()}.}
}
\value{
A ggplot object.
}
\description{
Draw a heatmap from a \code{\link[glyexp:experiment]{glyexp::experiment()}}.
Heatmap of log2-transformed expression values is plotted.
Before plotting, zero-variance rows and columns are filtered out
to ensure robust clustering.
}
