% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-logo.R
\name{plot_logo}
\alias{plot_logo}
\title{Logo Plot}
\usage{
plot_logo(exp, n_aa = 5L, fasta = NULL, ...)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{n_aa}{The number of amino acids to the left and right of the glycosylation site.
For example, if \code{n_aa = 5}, the resulting sequence will contain 11 amino acids.}

\item{fasta}{The path to the FASTA file containing protein sequences.
If \code{glyclean::add_site_seq()} has been called on the experiment,
this argument can be omitted.}

\item{...}{Additional arguments passed to \code{\link[ggseqlogo:ggseqlogo]{ggseqlogo::ggseqlogo()}}.}
}
\value{
A ggplot object.
}
\description{
This function accepts a \code{\link[glyexp:experiment]{glyexp::experiment()}},
and plots the logo plot for all glycosites.
Positions with insufficient flanking amino acids will be padded with "X".
}
\seealso{
\code{\link[ggseqlogo:ggseqlogo]{ggseqlogo::ggseqlogo()}}
}
