% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-logo.R
\name{plot_logo}
\alias{plot_logo}
\alias{plot_logo.glyexp_experiment}
\title{Logo Plot}
\usage{
plot_logo(x, n_aa = 5L, fasta = NULL, ...)

\method{plot_logo}{glyexp_experiment}(x, n_aa = 5L, fasta = NULL, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{n_aa}{The number of amino acids to the left and right of the glycosylation site.
For example, if \code{n_aa = 5}, the resulting sequence will contain 11 amino acids.}

\item{fasta}{The path to the FASTA file containing protein sequences.
If \code{glyclean::add_site_seq()} has been called on the experiment,
this argument can be omitted.}

\item{...}{Additional arguments passed to \code{\link[ggseqlogo:ggseqlogo]{ggseqlogo::ggseqlogo()}}.}
}
\value{
A ggplot object.
}
\description{
Draw a logo plot for all glycosites.
Positions with insufficient flanking amino acids will be padded with "X".
Currently supported data types:
\itemize{
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
Logo plot of glycosylation sites is plotted.
}
}
\seealso{
\code{\link[ggseqlogo:ggseqlogo]{ggseqlogo::ggseqlogo()}}
}
