% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pca.R
\name{plot_pca}
\alias{plot_pca}
\alias{plot_pca.glystats_pca_res}
\alias{plot_pca.glyexp_experiment}
\title{PCA Plot}
\usage{
plot_pca(x, type = "individual", ...)

\method{plot_pca}{glystats_pca_res}(x, type = "individual", ...)

\method{plot_pca}{glyexp_experiment}(x, type = "individual", groups = NULL, group_col = NULL, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{type}{The type of plot, one of "screeplot", "individual", "variables", or "biplot".}

\item{...}{Additional arguments passed to underlying factoextra functions:
\itemize{
\item type = "screeplot": \code{\link[factoextra:eigenvalue]{factoextra::fviz_screeplot()}}.
\item type = "individual": \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_ind()}}.
\item type = "variables": \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_var()}}.
\item type = "biplot": \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_biplot()}}.
}}

\item{groups}{A factor or character vector specifying group membership for each sample.
If provided, the plot will be colored by group.
Only applicable to "individual" and "biplot" types.
Passed to the \code{habillage} paramter of \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_ind()}} or \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_biplot()}}.}

\item{group_col}{A character string specifying where to find the group information.
If you uses \code{\link[glystats:gly_pca]{glystats::gly_pca()}} on a \code{\link[glyexp:experiment]{glyexp::experiment()}} to get the result,
sample information has already been added to the result.
In this case, you can specify the column name in the sample information tibble
to be used for coloring.
If not provided, this function will try "group".}
}
\value{
A ggplot object.
}
\description{
Draw a PCA scores plot.
Currently supported data types:
\itemize{
\item \code{glystats_pca_res}: Result from \code{\link[glystats:gly_pca]{glystats::gly_pca()}}.
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
PCA analysis is first performed using \code{\link[glystats:gly_pca]{glystats::gly_pca()}},
then the result is plotted.
}
}
\seealso{
\code{\link[factoextra:eigenvalue]{factoextra::fviz_screeplot()}}, \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_ind()}},
\code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_var()}}, \code{\link[factoextra:fviz_pca]{factoextra::fviz_pca_biplot()}}
}
