% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roc.R
\name{plot_roc}
\alias{plot_roc}
\alias{plot_roc.glystats_roc_res}
\alias{plot_roc.glyexp_experiment}
\title{ROC plot}
\usage{
plot_roc(x, type = "roc", auc_cutoff = 0.5, ...)

\method{plot_roc}{glystats_roc_res}(x, type = "roc", auc_cutoff = 0.5, ...)

\method{plot_roc}{glyexp_experiment}(x, type = "roc", auc_cutoff = 0.5, stats_args = list(), ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{type}{The type of plot, one of "dotplot" or "roc". Default is "roc".}

\item{auc_cutoff}{The AUC cutoff. Default is 0.5.}

\item{...}{Ignored.}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_roc]{glystats::gly_roc()}}.}
}
\value{
A ggplot object.
}
\description{
Draw ROC curves.
Only two groups are allowed.
At most 10 variables can be plotted.
Currently supported data types:
\itemize{
\item \code{glystats_roc_res}: Result from \code{\link[glystats:gly_roc]{glystats::gly_roc()}}.
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
ROC analysis is first performed using \code{\link[glystats:gly_roc]{glystats::gly_roc()}},
then the result is plotted.
}
}
