% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roc.R
\name{plot_roc}
\alias{plot_roc}
\alias{plot_roc.glyexp_experiment}
\title{ROC plot}
\usage{
plot_roc(x, type = "roc", auc_cutoff = 0.5, ...)

\method{plot_roc}{glyexp_experiment}(x, type = "roc", auc_cutoff = 0.5, group_col = "group", ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{type}{The type of plot, one of "dotplot" or "roc". Default is "roc".}

\item{auc_cutoff}{The AUC cutoff. Default is 0.5.}

\item{...}{Ignored.}

\item{group_col}{A character string specifying the column name of the grouping variable.
Defaults to "group".}
}
\value{
A ggplot object.
}
\description{
Draw ROC curves.
Only two groups are allowed.
At most 10 variables can be plotted.
}
