% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-umap.R
\name{plot_umap}
\alias{plot_umap}
\title{UMAP Plot}
\usage{
plot_umap(exp, ...)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{...}{Additional arguments passed to \code{\link[glystats:gly_umap]{glystats::gly_umap()}}.}
}
\value{
A ggplot object.
}
\description{
This function accepts a \code{\link[glyexp:experiment]{glyexp::experiment()}},
performs UMAP using \code{\link[glystats:gly_umap]{glystats::gly_umap()}},
and plots the UMAP scores plot.
}
