% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(exp, group_col = "group", p_adj_method = "BH", ref_group = NULL)
}
\arguments{
\item{exp}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}

\item{group_col}{A character string specifying the column name in sample information
that contains group labels. Default is "group".}

\item{p_adj_method}{A character string specifying the method to adjust p-values.
See \code{p.adjust.methods} for available methods. Default is "BH".
If NULL, no adjustment is performed.}

\item{ref_group}{A character string specifying the reference group.
If NULL (default), the first level of the group factor is used as the reference.}
}
\value{
A ggplot object.
}
\description{
This function accepts a \code{\link[glyexp:experiment]{glyexp::experiment()}},
performs differential expression analysis using \code{\link[glystats:gly_limma]{glystats::gly_limma()}},
and plots a volcano plot.
There should be only two groups.
}
