% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\alias{plot_volcano.glyexp_experiment}
\alias{plot_volcano.glystats_ttest_res}
\alias{plot_volcano.glystats_wilcox_res}
\alias{plot_volcano.glystats_limma_res}
\title{Volcano Plot}
\usage{
plot_volcano(x, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

\method{plot_volcano}{glyexp_experiment}(
  x,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  test = "limma",
  stats_args = list(),
  ...
)

\method{plot_volcano}{glystats_ttest_res}(x, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

\method{plot_volcano}{glystats_wilcox_res}(x, log2fc_cutoff = 1, p_cutoff = 0.05, p_col = "p_adj", ...)

\method{plot_volcano}{glystats_limma_res}(
  x,
  contrast = NULL,
  log2fc_cutoff = 1,
  p_cutoff = 0.05,
  p_col = "p_adj",
  ...
)
}
\arguments{
\item{x}{An object to be plotted.}

\item{log2fc_cutoff}{The log2 fold change cutoff. Defaults to 1.}

\item{p_cutoff}{The p-value cutoff. Defaults to 0.05.}

\item{p_col}{The column name for p-value. Defaults to "p_adj".
Can also be "p_val" (raw p-values without multiple testing correction).}

\item{...}{Other arguments passed to \code{\link[EnhancedVolcano:EnhancedVolcano]{EnhancedVolcano::EnhancedVolcano()}}.}

\item{test}{"ttest", "wilcox", or "limma". Default is "limma".}

\item{stats_args}{A list of keyword arguments to pass to \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}},
\code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}}, or \code{\link[glystats:gly_limma]{glystats::gly_limma()}}.}

\item{contrast}{A character string specifying the contrast to plot, in the format of "group1_vs_group2".
Must be one of the contrasts in the result.
When there is only one contrast (two-group comparison), it can be NULL (default).}
}
\value{
A ggplot object.
}
\description{
Draw a volcano plot.
Currently supported data types:
\itemize{
\item \code{glystats_ttest_res}: Result from \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}}.
\item \code{glystats_wilcox_res}: Result from \code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}}.
\item \code{glystats_limma_res}: Result from \code{\link[glystats:gly_limma]{glystats::gly_limma()}}.
Only two-group comparisons are supported.
\item \code{glyexp_experiment}: Experiment created by \code{\link[glyexp:experiment]{glyexp::experiment()}}.
Differential expression analysis is first performed using \code{\link[glystats:gly_ttest]{glystats::gly_ttest()}},
\code{\link[glystats:gly_wilcox]{glystats::gly_wilcox()}}, or \code{\link[glystats:gly_limma]{glystats::gly_limma()}}
(controlled by the \code{test} argument), then the result is plotted.
}
}
